+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Removed theme settings when updating page metadata: Saving the page settings, remove the theme selected when it is selected from the global or specific theme of a section #237
 
+### Updated
+
+- Public sections support for section theme components: #239
+
 ### [3.0.12] - 2025-09-09
 
 ### Fixed